Add properties resource (via scaffold)
This is the basis for property listings. It's used internally by the
agency to show properties around the agency and share key information.

There's no auth for now, because all this information is public and is
theoretically allowed to be shared by agents, though in practice we
rarely show these pages to buyers.